streamlit/requirements.txt
neo4j
streamlit
spotipy
lyricsgenius
Pillow
openai